RD-1853 Move table-related files to a separate directory
