+++ widgets/deploymentsView/src/table/index.ts
@@ -1,0 +1,4 @@
+import DeploymentsTable from './DeploymentsTable';
+
+export * from './columns';
+export { DeploymentsTable };
